AC-761 Перевел тесты за исключением истории

+++ app/src/main/java/com/obsudim/mypsychologist/domain/entity/diagnosticEntity/TestsEntity.kt
@@ -58,8 +58,8 @@ data class TestInfoEntity(
 
 data class BorderOfTestEntity(
     val color: String,
-    val leftBorder: Int,
-    val rightBorder: Int,
+    val leftBorder: Double,
+    val rightBorder: Double,
     val title: String
 )
